Add files via upload
Newest set of current meter designs. The tapered shaft is for connecting the 15mm coupler to the fin mount. The tolerances need to be refined. The new bottom bearing press has a longer shaft to help press the bearings into the coupler. I have tested the fins yet so not sure if they are good. The fins are to be cut from 2mm or .08" material (I am going to use acrylic